% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert}
\alias{assert}
\title{assert a condition and throw error if it is not TRUE}
\usage{
assert(con, error, trace = NULL)
}
\arguments{
\item{con}{condition to be asserted}

\item{error}{error message to be thrown}

\item{trace}{call chain to be printed in the error message, if not specified,
the call chain will be parsed from \code{sys.calls()} inside this function.}
}
\description{
assert a condition and throw error if it is not TRUE
}
