% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_vector}
\alias{check_vector}
\title{check if an object is a vector, and not a list, and throw error if not}
\usage{
check_vector(obj, error = "Object must be a vector!", trace = NULL)
}
\arguments{
\item{obj}{object to be checked}

\item{error}{error message}

\item{trace}{call chain to be printed in the error message, if not specified,
the call chain will be parsed from \code{sys.calls()} inside this function.}
}
\description{
check if an object is a vector, and not a list, and throw error if not
}
