% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cm2inch}
\alias{cm2inch}
\title{convert centimeter to inch}
\usage{
cm2inch(cm)
}
\arguments{
\item{cm}{a numeric vector of centimeter values}
}
\value{
a numeric vector of inch values
}
\description{
convert centimeter to inch
}
