% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_map}
\alias{color_map}
\title{Map the color values to the group names, auto rep to Extend color length}
\usage{
color_map(color, group)
}
\arguments{
\item{color}{vector of color value,
like: c('red', 'blue') or c('#F8766D', '#00BFC2')}

\item{group}{vector of group name}
}
\value{
color_map a named vector with group name as names
and color value as values
}
\description{
Map the color values to the group names, auto rep to Extend color length
}
