% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{creat_dir}
\alias{creat_dir}
\title{create a directory, if it does not exist, have no warning}
\usage{
creat_dir(path)
}
\arguments{
\item{path}{a character string specifying the path to a directory}
}
\value{
the path to the directory
}
\description{
create a directory, if it does not exist, have no warning
}
