% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_file_dir}
\alias{create_file_dir}
\title{create a directory for a file, if it does not exist}
\usage{
create_file_dir(file)
}
\arguments{
\item{file}{a character string specifying the path to a file}
}
\value{
the path to the directory
}
\description{
create a directory for a file, if it does not exist
}
