% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{file_lock}
\alias{file_lock}
\title{A safe file lock, auto unlock}
\usage{
file_lock(path, FUN, ..., exclusive = TRUE, timeout = 5000)
}
\arguments{
\item{path}{file path}

\item{FUN}{function to be applied, first argument is the file path}

\item{...}{other arguments passed to \code{FUN}}

\item{exclusive}{whether to lock exclusively, default is \code{TRUE}, passed to \code{\link[filelock]{lock}}}

\item{timeout}{lock timeout in milliseconds, default is \code{5000}, passed to \code{\link[filelock]{lock}}}
}
\value{
the result of \code{FUN}
}
\description{
A safe file lock, auto unlock
}
