% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_vector}
\alias{is_vector}
\title{judge if an object is a vector, and not a list}
\usage{
is_vector(obj)
}
\arguments{
\item{obj}{object to be checked}
}
\value{
\code{TRUE} if it is a vector, \code{FALSE} otherwise
}
\description{
judge if an object is a vector, and not a list
}
