% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_call_chain}
\alias{parse_call_chain}
\title{format call chain}
\usage{
parse_call_chain(trace = sys.calls())
}
\arguments{
\item{trace}{call chain, a list of call strings,created by \code{sys.calls()}}
}
\value{
\code{character(1)} formatted string
}
\description{
format call chain
}
